@@ -1,6 +1,10 @@
-import React from 'react';
-import { Dialog, DialogTitle, DialogContent, Divider } from '@material-ui/core';
-import { Alert } from '@material-ui/lab';
+import {
+  Alert,
+  Dialog,
+  DialogContent,
+  DialogTitle,
+  Divider,
+} from '@mui/material';
 import AddPatientForm, { PatientFormValues } from './AddPatientForm';
 
 interface Props {
