@@ -1,15 +1,13 @@
 /* eslint-disable no-void */
-import { Button, Container, Divider, Typography } from '@material-ui/core';
+import { Button, Container, Divider, Typography } from '@mui/material';
 import axios from 'axios';
 import React from 'react';
 import { BrowserRouter as Router, Link, Route, Routes } from 'react-router-dom';
-
 import { apiBaseUrl } from './constants';
-import { useStateValue } from './state';
-import { Patient } from './types';
-
 import PatientListPage from './pages/PatientListPage';
 import PatientPage from './pages/PatientPage';
+import { useStateValue } from './state';
+import { Patient } from './types';
 
 const App = () => {
   const [, dispatch] = useStateValue();
